requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
black==23.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.3
cryptography==45.0.3
dotenv==0.9.9
eval_type_backport==0.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
mistralai==1.8.1
mypy-extensions==1.0.0
packaging==23.1
paramiko==3.5.1
pathspec==0.11.1
platformdirs==3.8.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
